requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.84.0
pyodbc==5.0.1
sqlalchemy==2.0.25
python-dotenv==1.0.0
pydantic==2.5.3